compatibility issue

+++ luh/feature_extractors/token_probabilities.py
@@ -33,8 +33,8 @@ class FeatureExtractorTokenProbabilities(FeatureExtractorBase):
         :return: Token-level probabilities (sequence_length x vocab_size).
         """
         mask = logits.sum(dim=-1) != 0.
-        top_probas = torch.zeros(*logits.shape[:-1], self.top_n, device=logits.device, dtype=torch.float32)
-
+        # top_probas = torch.zeros(*logits.shape[:-1], self.top_n, device=logits.device, dtype=torch.float32)
+        top_probas = torch.zeros(*logits.shape[:-1], self.top_n, device=logits.device, dtype=logits.dtype)
         def get_top_n_probas(inpt_logits):
             res_p = F.softmax(inpt_logits / self.temperature, dim=-1)
             res_p = res_p.topk(self.top_n, dim=-1)[0]
